Consolidate bignum to just have gmp option since way faster than openssl, remove unused files
